+++ frontend/src/metabase/collections/components/CollectionHeader/CollectionHeader.tsx
@@ -3,9 +3,9 @@ import type { Location } from "history";
 
 import { Collection, CollectionId } from "metabase-types/api";
 
+import { CollectionMenu } from "../CollectionMenu";
 import CollectionCaption from "./CollectionCaption";
 import CollectionBookmark from "./CollectionBookmark";
-import CollectionMenu from "./CollectionMenu";
 import CollectionTimeline from "./CollectionTimeline";
 import { CollectionUpload } from "./CollectionUpload";
