+++ frontend/src/metabase/collections/components/CollectionMenu/index.ts
@@ -1,0 +1,1 @@
+export { CollectionMenu } from "./CollectionMenu";
